Simplify the public list dags endpoint
The endpoint that is used by the UI needs to be able to sort dags by the state of the last dag run which is a bit of an idiosyncratic thing to do.  Moving it to the internal API makes it not subject to backcompat and thus easier to change as needed.  And keeping the public one simpler is a good thing.
