+++ airflow-core/src/airflow/api_fastapi/core_api/routes/public/dags.py
@@ -22,33 +22,25 @@ from typing import Annotated
 from fastapi import Depends, HTTPException, Query, Response, status
 from fastapi.exceptions import RequestValidationError
 from pydantic import ValidationError
-from sqlalchemy import func, null, select, update
+from sqlalchemy import select, update
 
 from airflow.api.common import delete_dag as delete_dag_module
 from airflow.api_fastapi.common.dagbag import DagBagDep
 from airflow.api_fastapi.common.db.common import (
     SessionDep,
-    apply_filters_to_select,
     paginated_select,
 )
 from airflow.api_fastapi.common.parameters import (
-    FilterOptionEnum,
-    FilterParam,
     QueryDagDisplayNamePatternSearch,
     QueryDagIdPatternSearch,
     QueryDagIdPatternSearchWithNone,
     QueryExcludeStaleFilter,
-    QueryLastDagRunStateFilter,
     QueryLimit,
     QueryOffset,
     QueryOwnersFilter,
     QueryPausedFilter,
     QueryTagsFilter,
-    RangeFilter,
     SortParam,
-    _transform_dag_run_states,
-    datetime_range_filter_factory,
-    filter_param_factory,
 )
 from airflow.api_fastapi.common.router import AirflowRouter
 from airflow.api_fastapi.core_api.datamodels.dags import (
@@ -66,7 +58,6 @@ from airflow.api_fastapi.core_api.security import (
 from airflow.api_fastapi.logging.decorators import action_logging
 from airflow.exceptions import AirflowException, DagNotFound
 from airflow.models import DAG, DagModel
-from airflow.models.dagrun import DagRun
 
 dags_router = AirflowRouter(tags=["DAG"], prefix="/dags")
 
@@ -81,33 +72,12 @@ def get_dags(
     dag_display_name_pattern: QueryDagDisplayNamePatternSearch,
     exclude_stale: QueryExcludeStaleFilter,
     paused: QueryPausedFilter,
-    last_dag_run_state: QueryLastDagRunStateFilter,
-    dag_run_start_date_range: Annotated[
-        RangeFilter, Depends(datetime_range_filter_factory("dag_run_start_date", DagRun, "start_date"))
-    ],
-    dag_run_end_date_range: Annotated[
-        RangeFilter, Depends(datetime_range_filter_factory("dag_run_end_date", DagRun, "end_date"))
-    ],
-    dag_run_state: Annotated[
-        FilterParam[list[str]],
-        Depends(
-            filter_param_factory(
-                DagRun.state,
-                list[str],
-                FilterOptionEnum.ANY_EQUAL,
-                "dag_run_state",
-                default_factory=list,
-                transform_callable=_transform_dag_run_states,
-            )
-        ),
-    ],
     order_by: Annotated[
         SortParam,
         Depends(
             SortParam(
                 ["dag_id", "dag_display_name", "next_dagrun", "state", "start_date"],
                 DagModel,
-                {"last_run_state": DagRun.state, "last_run_start_date": DagRun.start_date},
             ).dynamic_depends()
         ),
     ],
@@ -115,47 +85,8 @@ def get_dags(
     session: SessionDep,
 ) -> DAGCollectionResponse:
     """Get all DAGs."""
-    query = select(DagModel)
-
-    max_run_id_query = (  # ordering by id will not always be "latest run", but it's a simplifying assumption
-        select(DagRun.dag_id, func.max(DagRun.id).label("max_dag_run_id"))
-        .where(DagRun.start_date.is_not(null()))
-        .group_by(DagRun.dag_id)
-        .subquery(name="mrq")
-    )
-
-    has_max_run_filter = (
-        dag_run_state.value
-        or last_dag_run_state.value
-        or dag_run_start_date_range.is_active()
-        or dag_run_end_date_range.is_active()
-    )
-
-    if has_max_run_filter or order_by.value in (
-        "last_run_state",
-        "last_run_start_date",
-        "-last_run_state",
-        "-last_run_start_date",
-    ):
-        query = query.join(
-            max_run_id_query,
-            DagModel.dag_id == max_run_id_query.c.dag_id,
-            isouter=True,
-        ).join(DagRun, DagRun.id == max_run_id_query.c.max_dag_run_id, isouter=True)
-
-    if has_max_run_filter:
-        query = apply_filters_to_select(
-            statement=query,
-            filters=[
-                dag_run_start_date_range,
-                dag_run_end_date_range,
-                dag_run_state,
-                last_dag_run_state,
-            ],
-        )
-
     dags_select, total_entries = paginated_select(
-        statement=query,
+        statement=select(DagModel),
         filters=[
             exclude_stale,
             paused,
